+++ package/lib/src/easy_layout_auto.dart
@@ -1,6 +1,20 @@
 import 'package:easy_layout/easy_layout.dart';
 import 'package:flutter/widgets.dart';
 
+/// A widget that displays its children in a horizontal or
+/// vertical array like a [Row] or [Column] widgets,
+/// but with spacing and change the horizontal to vertical
+/// layout when changing the screen width.
+///
+/// In a horizontal layout, child widgets are expanded.
+///
+/// By default inherits spacing from parent [EasyLayout].
+///
+/// The `toggleWidth` parameter specifies the minimum
+/// width for a horizontal layout, if the width is
+/// less than the specified one, the layout will switch
+/// to a vertical layout.
+///
 class EasyLayoutAuto extends StatelessWidget {
   final List<Widget> children;
   final double hSpacing;