+++ package/lib/src/easy_layout_flexible.dart
@@ -1,5 +1,11 @@
 import 'package:flutter/widgets.dart';
 
+/// A widget that controls how a child of a EasyLayoutRow or EasyLayoutAuto flexes.
+///
+/// See also:
+/// * [Flexible], which works similarly for [Column], [Row], and [Flex].
+/// * [Space distribution explanation](https://pub.dev/packages/easy_layout#space-distribution)
+///
 class EasyLayoutFlexible extends StatelessWidget {
   final Widget child;
   final double flex;