+++ package/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+* EasyLayout now inherits spacing from parent EasyLayout
+* Added EasyLayoutSpacing widget
+* Added EasyLayoutFluid widget
+
 ## 0.1.0
 
 * Initial version